pyproject.toml
[project]
name = "icnns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Akishika, Issui", email = "kshkss@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dm-haiku>=0.0.15",
    "jax>=0.8.1",
]

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[tool.pyright]
venvPath = "."
venv = ".venv"
